ROFO-15 refactor: 패키지 분리
