+++ data/src/main/java/com/weit2nd/data/di/LoginModule.kt
@@ -1,9 +1,9 @@
 package com.weit2nd.data.di
 
-import com.weit2nd.data.repository.LoginRepositoryImpl
-import com.weit2nd.data.source.LoginDataSource
+import com.weit2nd.data.repository.login.LoginRepositoryImpl
+import com.weit2nd.data.source.login.LoginDataSource
 import com.weit2nd.data.util.ActivityProvider
-import com.weit2nd.domain.repository.LoginRepository
+import com.weit2nd.domain.repository.login.LoginRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
