+++ domain/src/main/java/com/weit2nd/domain/repository/login/LoginRepository.kt
@@ -1,4 +1,4 @@
-package com.weit2nd.domain.repository
+package com.weit2nd.domain.repository.login
 
 interface LoginRepository {
     suspend fun loginWithKakao(): Result<Unit>
